Move OakCamera from __init__.py to oak_camera.py, refactor formatting
